Check for nix-env 1.7 in bin/shell before starting.
Summary:
Things go horribly, horribly wrong when using nix-env 1.6 instead of 1.7.
Ideally, this should probably be checked in nix expressions themselves, but
since this is our main entry point, it seems reasonable to add it here for now.

Test Plan:
Changed the grep to '1.6' instead of '1.7' to test, and got the error as
expected.

Reviewers: johnw, austin

Reviewed By: austin

Differential Revision: https://phabricator.haskell.org/D9